update App, home, create auth/appscreens
