+++ client/auth_screens/AuthScreens.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function AuthScreens() {
+  return (
+    <div>AuthScreens</div>
+  );
+};
